resource/sns_topic_subscription: Move obfuscation to test
